UPD: remove useless methods et attributes (estimation distribution) for classification score
