+++ mapie/conformity_scores/interface.py
@@ -3,7 +3,6 @@ from abc import ABCMeta, abstractmethod
 import numpy as np
 
 from mapie._compatibility import np_nanquantile
-from mapie._machine_precision import EPSILON
 from mapie._typing import NDArray
 
 
@@ -12,34 +11,10 @@ class BaseConformityScore(metaclass=ABCMeta):
     Base class for conformity scores.
 
     This class should not be used directly. Use derived classes instead.
-
-    Parameters
-    ----------
-    consistency_check: bool, optional
-        Whether to check the consistency between the methods
-        ``get_estimation_distribution`` and ``get_conformity_scores``.
-        If ``True``, the following equality must be verified:
-        ``self.get_estimation_distribution(
-            y_pred, self.get_conformity_scores(y, y_pred, **kwargs), **kwargs
-        ) == y``
-
-        By default ``True``.
-
-    eps: float, optional
-        Threshold to consider when checking the consistency between
-        ``get_estimation_distribution`` and ``get_conformity_scores``.
-        It should be specified if ``consistency_check==True``.
-
-        By default, it is defined by the default precision.
     """
 
-    def __init__(
-        self,
-        consistency_check: bool = True,
-        eps: float = float(EPSILON),
-    ):
-        self.consistency_check = consistency_check
-        self.eps = eps
+    def __init__(self) -> None:
+        pass
 
     def set_external_attributes(
         self,
@@ -53,56 +28,6 @@ class BaseConformityScore(metaclass=ABCMeta):
         instantiated.
         """
         pass
-
-    def check_consistency(
-        self,
-        y: NDArray,
-        y_pred: NDArray,
-        conformity_scores: NDArray,
-        **kwargs
-    ) -> None:
-        """
-        Check consistency between the following methods:
-        ``get_estimation_distribution`` and ``get_signed_conformity_scores``
-
-        The following equality should be verified:
-        ``self.get_estimation_distribution(
-            y_pred, self.get_conformity_scores(y, y_pred, **kwargs), **kwargs
-        ) == y``
-
-        Parameters
-        ----------
-        y: NDArray of shape (n_samples, ...)
-            Observed target values.
-
-        y_pred: NDArray of shape (n_samples, ...)
-            Predicted target values.
-
-        conformity_scores: NDArray of shape (n_samples, ...)
-            Conformity scores.
-
-        Raises
-        ------
-        ValueError
-            If the two methods are not consistent.
-        """
-        score_distribution = self.get_estimation_distribution(
-            y_pred, conformity_scores, **kwargs
-        )
-        abs_conformity_scores = np.abs(np.subtract(score_distribution, y))
-        max_conf_score = np.max(abs_conformity_scores)
-        if max_conf_score > self.eps:
-            raise ValueError(
-                "The two functions get_conformity_scores and "
-                "get_estimation_distribution of the BaseConformityScore class "
-                "are not consistent. "
-                "The following equation must be verified: "
-                "self.get_estimation_distribution(y_pred, "
-                "self.get_conformity_scores(y, y_pred)) == y. "
-                f"The maximum conformity score is {max_conf_score}. "
-                "The eps attribute may need to be increased if you are "
-                "sure that the two methods are consistent."
-            )
 
     @abstractmethod
     def get_conformity_scores(
@@ -130,34 +55,6 @@ class BaseConformityScore(metaclass=ABCMeta):
         -------
         NDArray of shape (n_samples, ...)
             Conformity scores.
-        """
-
-    @abstractmethod
-    def get_estimation_distribution(
-        self,
-        y_pred: NDArray,
-        conformity_scores: NDArray,
-        **kwargs
-    ) -> NDArray:
-        """
-        Placeholder for ``get_estimation_distribution``.
-        Subclasses should implement this method!
-
-        Compute samples of the estimation distribution given the predicted
-        targets and the conformity scores.
-
-        Parameters
-        ----------
-        y_pred: NDArray of shape (n_samples, ...)
-            Predicted target values.
-
-        conformity_scores: NDArray of shape (n_samples, ...)
-            Conformity scores.
-
-        Returns
-        -------
-        NDArray of shape (n_samples, ...)
-            Observed values.
         """
 
     @staticmethod
