+++ mapie/conformity_scores/regression.py
@@ -41,12 +41,15 @@ class BaseRegressionScore(BaseConformityScore, metaclass=ABCMeta):
     """
 
     def __init__(
-        self, sym: bool,
+        self,
+        sym: bool,
         consistency_check: bool = True,
         eps: float = float(EPSILON),
     ):
-        super().__init__(consistency_check=consistency_check, eps=eps)
+        super().__init__()
         self.sym = sym
+        self.consistency_check = consistency_check
+        self.eps = eps
 
     @abstractmethod
     def get_signed_conformity_scores(
@@ -105,6 +108,84 @@ class BaseRegressionScore(BaseConformityScore, metaclass=ABCMeta):
         if self.sym:
             conformity_scores = np.abs(conformity_scores)
         return conformity_scores
+
+    def check_consistency(
+        self,
+        y: NDArray,
+        y_pred: NDArray,
+        conformity_scores: NDArray,
+        **kwargs
+    ) -> None:
+        """
+        Check consistency between the following methods:
+        ``get_estimation_distribution`` and ``get_signed_conformity_scores``
+
+        The following equality should be verified:
+        ``self.get_estimation_distribution(
+            y_pred, self.get_conformity_scores(y, y_pred, **kwargs), **kwargs
+        ) == y``
+
+        Parameters
+        ----------
+        y: NDArray of shape (n_samples, ...)
+            Observed target values.
+
+        y_pred: NDArray of shape (n_samples, ...)
+            Predicted target values.
+
+        conformity_scores: NDArray of shape (n_samples, ...)
+            Conformity scores.
+
+        Raises
+        ------
+        ValueError
+            If the two methods are not consistent.
+        """
+        score_distribution = self.get_estimation_distribution(
+            y_pred, conformity_scores, **kwargs
+        )
+        abs_conformity_scores = np.abs(np.subtract(score_distribution, y))
+        max_conf_score = np.max(abs_conformity_scores)
+        if max_conf_score > self.eps:
+            raise ValueError(
+                "The two functions get_conformity_scores and "
+                "get_estimation_distribution of the BaseConformityScore class "
+                "are not consistent. "
+                "The following equation must be verified: "
+                "self.get_estimation_distribution(y_pred, "
+                "self.get_conformity_scores(y, y_pred)) == y. "
+                f"The maximum conformity score is {max_conf_score}. "
+                "The eps attribute may need to be increased if you are "
+                "sure that the two methods are consistent."
+            )
+
+    @abstractmethod
+    def get_estimation_distribution(
+        self,
+        y_pred: NDArray,
+        conformity_scores: NDArray,
+        **kwargs
+    ) -> NDArray:
+        """
+        Placeholder for ``get_estimation_distribution``.
+        Subclasses should implement this method!
+
+        Compute samples of the estimation distribution given the predicted
+        targets and the conformity scores.
+
+        Parameters
+        ----------
+        y_pred: NDArray of shape (n_samples, ...)
+            Predicted target values.
+
+        conformity_scores: NDArray of shape (n_samples, ...)
+            Conformity scores.
+
+        Returns
+        -------
+        NDArray of shape (n_samples, ...)
+            Observed values.
+        """
 
     @staticmethod
     def _beta_optimize(
